+++ open_mafia_engine/built_in/load.py
@@ -4,11 +4,11 @@ __all__ = ["prefabs"]
 
 from pathlib import Path
 
-from open_mafia_engine.core.ability import Ability
-from open_mafia_engine.core.constraint import Constraint
-from open_mafia_engine.core.phase import Phase
-from open_mafia_engine.core.prefab import Prefab
-from open_mafia_engine.core.wincon import Wincon
+from open_mafia_engine.state.ability import Ability
+from open_mafia_engine.state.constraint import Constraint
+from open_mafia_engine.state.phase import Phase
+from open_mafia_engine.state.prefab import Prefab
+from open_mafia_engine.state.wincon import Wincon
 from open_mafia_engine.util.yaml import load_yaml
 
 _p = Path(__file__).parent
