+++ open_mafia_engine/state/alignment.py
