+++ open_mafia_engine/state/prefab.py
