+++ open_mafia_engine/state/role.py
@@ -1,5 +1,5 @@
 from typing import List
-from open_mafia_engine.core.ability import Ability
+from open_mafia_engine.state.ability import Ability
 from pydantic import BaseModel, validator
