+++ open_mafia_engine/state/wincon.py
